Add files to handle, display, and create Users. Added functionality to Settings
I created a set of files to handle Users in the same way that we are
handling Chores and Resources, this allows us to display them as a list
and handle between activities.

Added listeners to the buttons on the Settings screen, added some
functionality.

Clicking LogOut now takes you to a list of existing users.

Added screen to create individual user. Can be used as a template for
editing User as well.